Merge nodes with Vegsperring with corresponding way nodes.
Unfortunately, the nodes with OBJTYPE=Vegsperring are sometimes
not on the same loacation as a way node. They are then on a way
segment, but not connected to the way. This means that it may be
better to use the data from XXXXSperr.txt instead of the nodes
with OBJTYPE=Vegsperring.