Handle case where XXXXHoyde.txt does not exist.
If XXXXHoyde.txt (or more precisely the file in sys.argv[3]) is
missing, ignore it. Of course it would be cleaner to either not
take this as input (e.g. only directory and municipality number),
or specify only those files that are used. However, this way
is the least changes and works for now.