Write Elveg SOSI barriers that are not on a node to separate file.
The barriers are also included in the XXXXSperr.txt files with
link to road id (TRANSID), but those do not say anything about the
type of barrier. The barriers that do not share coordinats with a
way node are not written to the separate file
XXXXdetached_barriers.osm. They may be manually reviewed and merged
from there.