Remove UnicodeEncodeErrors by handling unicode/strings consistently.
The data from osmapis is unicode strings, so running s.format(var)
where s is a string and var is a unicode string will give a unicode
error if var contains non-ascii characters. The solution is to
have s as a unicode string as well.

The unicode strings are then decoded on output, most of them in
the warn() function. The decoding is hard coded to UTF-8. The
reason for this is that althogh Python will use the encoding from the
environment if sys.stderr/stdout goes to a tty. However, when the
output is redirected to a file, ascii is implicitly used, which causes
errors whenever there are non-ascii characters in the warning messages,
which may happen if the reported variables contain non-ascii
characters.