Merge all nodes with the same coordinates
The input file produced by sosi2osm often has several nodes in the
same location (both nodes that are parts of a way and others).
With this commit, all merging is moved to after deletion of objects.
In order to update the ways that have their nodes merged, each node
gets a set of way_ids that holds the ways that use the node. The
merging then keeps this set up-to-date.