Ignore road data for roads where VPA length is shorter than road data length
There are several roads in the data where the VPA length of a road
(the length given in the map data) is shorter than the end points
specified in the road data (*Fart.txt, *Hoyde.txt, etc.). This is
inconsistent, and previously the road data above the VPA length was
simply stripped off. Now, ignore any road data that is inconsistent.
This has the added benefit that for roads that have VPA length 0,
which would make the code crash before (in municipalities 1119 and
1149), the script will now continue and run to the end.