Delete ways with no tags or only nvdb:id tag
Only ways with tag with action=delete would be deleted, but also
ways that have no tags or only nvdb:id tags are useless in an import.

Deleted nodes and ways are (already) written to a separate file, where
their Elveg tags are shown, making it easier to make sense of why a way
or node was deleted.