Allow for different maxspeed:forward and maxspeed:backward
Interpret the lanes column in the speed- and height limit files,
and compare to the lanes field from the SOSI file. When they
differ, it is mostly due to inconsistensies and errors in the
data, but the case of reduced speed in one direction (for a
regular two-lane road) seems to be correct most of the time.

Quite a lot of cleanup is required after interpreting the
data literally. E.g. both maxspeed and maxspeed:forward, or
only maxspeed:forward occurs in the data. In the end either
only the maxspeed tag should be present, or both maxspeed:forward
and maxspeed:backward with different values.